Fixed signed/unsigned problem.

+++ Group.cpp
@@ -142,7 +142,7 @@ bool CGroup::ReadSettings(bool bIsRefresh)
 
 	// Positions
 	m_nMonitor	= LiteStep::GetPrefixedRCInt(m_szName, "Monitor", 0);
-	if (m_nMonitor >= g_Monitors.size())
+	if ((UINT)m_nMonitor >= g_Monitors.size())
 	{
 		utils::ErrorMessage(E_WARNING, "There is no monitor #%i", m_nMonitor);
 		m_nMonitor = 0;
